pyproject.toml
[project]
name = "rkbv4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chonkie[semantic]>=1.5.6",
    "chromadb>=1.5.2",
    "django>=6.0.3",
    "django-click>=2.5.0",
    "django-llm-chat",
    "django-ninja>=1.5.3",
    "django-stubs>=5.2.9",
    "httpx>=0.28.1",
    "litellm>=1.82.0",
    "model-bakery>=1.23.3",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "textual>=8.0.2",
]

[tool.uv.sources]
django-llm-chat = { git = "https://github.com/sadeh-congnition/django-llm-chat.git" }
